Implemented a weapon system
ShipObjects will now require a linked ShipWeaponSystem script, which will have a reference to all Weapon objects that are attached to the ship. The Weapon Script can be used as an interface to create new weapon types, like the LaserGun script.

So far these weapons only print to the log with a working cooldown mechanic. We will need to implement actual projectiles/bullets/lasers next and have them be created from the weapons.