Adjusted laser collision and mesh creation
The laser mesh creation is now sepperated from the polygon defining using the triangles array. The laser's ends now adjust is lenths depending on how long the laser is currently and if it has collided with an object.

Laser collision detection has also been slightly adjusted to be more accurate.